refactor(docs): fold description_cap extension into seo_context
description_cap.py existed only to re-register a priority-900 hook
from seo_context's regex stack. Cross-extension private import was a
smell; merging removes one entry from extensions=[...] and keeps the
hook-priority logic next to the regex it depends on.

Follow-up to C1 in docs_pipeline_architecture_review.md
